scrapers/requirements.txt
# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# AI/ML
openai==1.10.0

# Utilities
python-dateutil==2.8.2
pydantic==2.5.3
tenacity==8.2.3  # Retry logic
